fix commit

+++ src/app/bot/modules/main/router.py
@@ -5,6 +5,9 @@ from aiogram.types import Message, ReplyKeyboardRemove
 from aiogram.filters.state import StateFilter
 
 from app.bot.modules.main.settings import settings
+from core.logging.api import get_loggers
+from application.use_cases.db.create_user import GetOrCreateUser
+from infrastructure.db.uow import UnitOfWork
 
 
 router: Router = Router(name="main")
@@ -21,6 +24,20 @@ async def main(
 ) -> None:
     """Отправляет пользователю reply клавиатуру главного меню."""
 
+    name: str = message.from_user.first_name
+    telegram = message.from_user.id
+    logging_data = get_loggers(name=settings.NAME_FOR_LOG_FOLDER)
+
+    result = await GetOrCreateUser(uow=UnitOfWork(), logging_data=logging_data).execute(
+        name=name,
+        telegram=telegram,
+    )
+    if not result.ok:
+        await message.answer(
+            text=f"{result.error.message}\n\nПопробуйте нажать \start снова"
+        )
+        return
+
     # Удаляет сообщение которое было последним
     try:
         await bot.delete_message(
@@ -30,12 +47,6 @@ async def main(
         pass
 
     await message.answer(
-        text="☕ Будь как дома путник",
+        text=f"☕ Будь как дома, {name}",
         reply_markup=ReplyKeyboardRemove(),
-    )
-
-    await bot.send_photo(
-        chat_id=message.chat.id,
-        photo=settings.MENU_IMAGE_FILE_ID,
-        reply_markup=get_main_inline_keyboards,
     )
